refactor: CustomCouponDesign types 리팩토링
