+++ frontend/src/api/post.ts
@@ -1,8 +1,8 @@
 import { api } from '.';
-import { CouponSettingDto } from '../pages/Admin/CustomCouponDesign';
 import { StampFormData } from '../pages/Admin/EarnStamp';
 import { CafeFormData } from '../pages/Admin/RegisterCafe';
 import { PostIsFavoritesReq } from '../pages/CouponList';
+import { CouponSettingReq } from '../types';
 
 export const postEarnStamp = async ({ earningStampCount, customerId, couponId }: StampFormData) => {
   await api.post(`/admin/customers/${customerId}/coupons/${couponId}/stamps`, {
@@ -20,7 +20,7 @@ export const postIssueCoupon = async (customerId: string) => {
     .then((response) => response.json());
 };
 
-export const postCouponSetting = async (cafeId: number, couponConfig: CouponSettingDto) => {
+export const postCouponSetting = async (cafeId: number, couponConfig: CouponSettingReq) => {
   return await api.post(`/admin/coupon-setting?cafe-id=${cafeId}`, couponConfig);
 };
