+++ frontend/src/pages/Admin/CustomCouponDesign/style.tsx
@@ -10,30 +10,29 @@ export const ImageUploadContainer = styled.div`
   justify-content: space-between;
 `;
 
-export const CouponContainer = styled.div``;
-
 export const SaveButtonWrapper = styled.div`
   display: flex;
   flex-direction: row-reverse;
   width: 100%;
 `;
 
+// FIXME: 빈 스타일드 컴포넌트 수정 요망
 export const PreviewCouponContainer = styled.div``;
 
-interface StyledImageSizeProps {
+export const CouponContainer = styled.div``;
+
+export const PreviewImageWrapper = styled.div<{
   $width: number;
   $height: number;
   $opacity?: number;
-}
-
-export const PreviewImageWrapper = styled.div<StyledImageSizeProps>`
+}>`
   position: relative;
   width: ${({ $width }) => `${$width}px`};
   height: ${({ $height }) => `${$height}px`};
   border: 1px dotted ${({ theme }) => theme.colors.black};
 `;
 
-export const PreviewImage = styled.img<StyledImageSizeProps>`
+export const PreviewImage = styled.img<{ $width: number; $height: number; $opacity?: number }>`
   width: ${({ $width }) => `${$width}px`};
   height: ${({ $height }) => `${$height}px`};
   opacity: ${({ $opacity }) => ($opacity ? $opacity : '1')};
